Housekeeping work. Integration of new i18n.jar release.
git-svn-id: http://argouml.tigris.org/svn/argouml/trunk/src_new@5231 a161b567-7d1e-0410-9ef9-912c70fedb3f

+++ org/argouml/ui/ZoomSliderButton.java
@@ -179,7 +179,7 @@ public class ZoomSliderButton extends PopupButton {
         updateCurrentValueLabel();
                 
         //if (!source.getValueIsAdjusting()) {
-        double zoomPercentage = (double) _slider.getValue() / 100d;
+        double zoomPercentage = _slider.getValue() / 100d;
                 
         Editor ed = Globals.curEditor();
         if (ed == null || zoomPercentage <= 0.0) {
